refactor(assets): split index.ts into lib modules
Pure relocation of the 913-line assets plugin entry (grew past the SPEC
threshold with the #457 enrichment work) into lib modules, mirroring the
tasks/schedule/models thin-index template:

- lib/routes.ts — the 22 declarative routes + searchRoute as assetsRoutes
- lib/register-search.ts — the file-backed `assets` content-type
  registration (schema v2, text+visual legs, manifest-driven onSync/
  onUnlink, store-walk reindex) + the module-scope pluginCtx cell and
  indexVersionedAsset (one owning module, schedule's plugin-context
  precedent)
- lib/register-hooks.ts — the assets.* cross-plugin hooks
  (enrichmentStats/listByTask/getAssetTypes/resolveServe/saveFromSource/
  purgeClipboardForTask)
- lib/exec-tools.ts — all 14 bakin_exec_assets_* MCP tools + TYPE_RUBRIC

index.ts is now activate-orchestration + lifecycle (124 lines): enrichment
queue init, unmanaged emitter, health check, onReady/onShutdown. Bodies
moved byte-for-byte; hook names/payloads, tool names, route paths, and
lib/asset-service exports (the assets-bridge lazy-import surface) are
unchanged.
